Moving functions from __init__ to process under utils and updated imports
