+++ gprofiler/profilers/dotnet.py
@@ -18,8 +18,8 @@ from gprofiler.log import get_logger_adapter
 from gprofiler.metadata.application_metadata import ApplicationMetadata
 from gprofiler.profilers.profiler_base import ProcessProfilerBase
 from gprofiler.profilers.registry import register_profiler
-from gprofiler.utils import pgrep_maps, random_prefix, removed_path, resource_path, run_process
-from gprofiler.utils.process import is_process_basename_matching, process_comm
+from gprofiler.utils import  random_prefix, removed_path, resource_path
+from gprofiler.utils.process import pgrep_maps, is_process_basename_matching, process_comm, run_process
 from gprofiler.utils.speedscope import load_speedscope_as_collapsed
 
 logger = get_logger_adapter(__name__)
