+++ gprofiler/profilers/php.py
@@ -19,7 +19,8 @@ from gprofiler.gprofiler_types import ProcessToProfileData, ProcessToStackSample
 from gprofiler.log import get_logger_adapter
 from gprofiler.profilers.profiler_base import ProfilerBase
 from gprofiler.profilers.registry import ProfilerArgument, register_profiler
-from gprofiler.utils import random_prefix, resource_path, start_process, wait_event
+from gprofiler.utils import random_prefix, resource_path, wait_event
+from gprofiler.utils.process import start_process
 
 logger = get_logger_adapter(__name__)
 # Currently tracing only php-fpm, TODO: support mod_php in apache.
